vp9_firstpass: Fix MSVC data loss warning.

+++ vp9/encoder/vp9_firstpass.c
@@ -2402,8 +2402,8 @@ void vp9_twopass_postencode_update(VP9_COMP *cpi) {
   const double progress =
       (double)(cpi->twopass.stats_in - cpi->twopass.stats_in_start) /
               (cpi->twopass.stats_in_end - cpi->twopass.stats_in_start);
-  const int bits_used = progress * rc->this_frame_target +
-                        (1.0 - progress) * rc->projected_frame_size;
+  const int bits_used = (int)(progress * rc->this_frame_target +
+                             (1.0 - progress) * rc->projected_frame_size);
 #endif
 
   cpi->twopass.bits_left -= bits_used;
